create util lib to share common code across test apps
