+++ code/app-kotlin/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 
 <resources>
-    <string name="app_name">Edge Test App kt</string>
+    <string name="app_name">Edge Test App Kotlin</string>
 
 <!--    Edge view -->
     <string name="no_data">No data available</string>
