+++ code/app/build.gradle
@@ -61,6 +61,7 @@ spotless {
 dependencies {
     // Add dependencies on AEP Mobile extensions
     implementation project(':edge')
+    implementation project(':app-util-xdm')
 
     // TODO - Uncomment next lines when 2.0 is published to Maven
     //implementation 'com.adobe.marketing.mobile:core:2.+'
