+++ code/app/src/main/java/com/adobe/marketing/tester/MainActivity.java
@@ -37,11 +37,7 @@ import com.adobe.marketing.mobile.edge.identity.Identity;
 import com.adobe.marketing.mobile.edge.identity.IdentityItem;
 import com.adobe.marketing.mobile.edge.identity.IdentityMap;
 import com.adobe.marketing.mobile.services.Log;
-import com.adobe.marketing.mobile.xdm.Commerce;
-import com.adobe.marketing.mobile.xdm.MobileSDKCommerceSchema;
-import com.adobe.marketing.mobile.xdm.Order;
-import com.adobe.marketing.mobile.xdm.ProductListAdds;
-import com.adobe.marketing.mobile.xdm.Purchases;
+import com.adobe.marketing.tester.xdm.commerce.*;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import java.util.ArrayList;
